+++ packages/yourloops/lib/auth/models.ts
@@ -80,7 +80,7 @@ export interface AuthAPI {
   accountConfirmed: (key: string, traceToken: string) => Promise<boolean>;
   certifyProfessionalAccount: () => Promise<IUser>;
   getUserInfo: (session: Session) => Promise<User>;
-  getShorelineAccessToken: (email: string) => Promise<string>;
+  getShorelineAccessToken: (email: string) => Promise<[string, string?]>;
   refreshToken: (session: Readonly<Session>) => Promise<string>;
   resendSignup: (username: string, traceToken: string, language?: string) => Promise<boolean>;
   sendAccountValidation: (session: Readonly<Session>, language?: string) => Promise<boolean>;